App main method created
