+++ src/familybot/App.java
@@ -1,12 +1,22 @@
 package familybot;
 
+import java.awt.EventQueue;
+
+import UI.MainWindows;
+
 public class App {
 
-    /**
-     * @param args the command line arguments
-     */
     public static void main(String[] args) {
-        
+    	EventQueue.invokeLater(new Runnable() {
+			public void run() {
+				try {
+					MainWindows frame = new MainWindows();
+					frame.setVisible(true);
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
+			}
+		});
     }
     
 }
